+++ app/(tabs)/(home)/index.tsx
@@ -1,1 +1,6 @@
-export { default } from '../index';
+import React from 'react';
+import HomeScreen from '../index';
+
+export default function HomeTabIndex() {
+  return <HomeScreen />;
+}