+++ app/(tabs)/(styles)/index.tsx
@@ -1,1 +1,6 @@
-export { default } from './settings';
+import React from 'react';
+import SettingsScreen from './settings';
+
+export default function StylesTabIndex() {
+  return <SettingsScreen />;
+}